refactor(status): remove unreachable scan modes
