+++ src/commands/status.cold-imports.test.ts
@@ -43,7 +43,7 @@ describe("status cold imports", () => {
     await expect(resolveStatusRuntimeSnapshot(params)).resolves.toEqual(snapshot);
   });
 
-  it("keeps broad plugin status code behind the detailed status boundary", async () => {
+  it("keeps broad plugin status code out of default status imports", async () => {
     vi.doMock("../plugins/status.js", () => {
       throw new Error("default status must not import broad plugin diagnostics");
     });
