+++ src/commands/status.scan.fast-json.test.ts
@@ -306,6 +306,46 @@ describe("scanStatusJsonFast", () => {
     expect(mocks.probeGateway).not.toHaveBeenCalled();
   });
 
+  it("keeps status --json on read-only channel metadata when channel config exists", async () => {
+    const resolvedConfig = {
+      marker: "resolved-preload",
+      plugins: { enabled: false },
+      channels: { telegram: { enabled: false } },
+    };
+    applyStatusScanDefaults(mocks, {
+      hasConfiguredChannels: true,
+      sourceConfig: {
+        marker: "source-preload",
+        plugins: { enabled: false },
+        channels: { telegram: { enabled: false } },
+      } as never,
+      resolvedConfig: resolvedConfig as never,
+      summary: createStatusSummary({ linkChannel: { linked: false } }),
+    });
+
+    await scanStatusJsonFast({}, {} as never);
+
+    expect(mocks.ensurePluginRegistryLoaded).not.toHaveBeenCalled();
+    expect(loggingStateRef.forceConsoleToStderr).toBe(false);
+    expect(mocks.probeGateway).toHaveBeenCalledOnce();
+    const probeArgs = firstCallArg(mocks.probeGateway, "probeGateway args") as {
+      env?: NodeJS.ProcessEnv;
+    };
+    expect(probeArgs).toMatchObject({
+      url: "ws://127.0.0.1:18789",
+      config: resolvedConfig,
+      auth: {},
+      timeoutMs: 1000,
+      detailLevel: "presence",
+    });
+    expect(probeArgs.env).toBe(process.env);
+    expect(
+      mocks.callGateway.mock.calls.some(([call]) => {
+        return (call as { method?: unknown } | undefined)?.method === "channels.status";
+      }),
+    ).toBe(false);
+  });
+
   it("keeps cold-start gateway probes with local-only updates when a channel is configured from manifest env vars", async () => {
     await withTemporaryEnv(
       {
