+++ src/commands/status.scan.ts
@@ -1,66 +1,35 @@
-// Top-level status scan entrypoint.
-// Chooses fast JSON policy or full human scan and returns one normalized scan result.
+// Top-level text status scan entrypoint.
+// Human `status --all` and JSON status use their dedicated command paths.
 
 import { withProgress } from "../cli/progress.js";
-import { hasConfiguredChannelsForReadOnlyScope } from "../plugins/channel-plugin-ids.js";
-import type { RuntimeEnv } from "../runtime.js";
 import { executeStatusScanFromOverview } from "./status.scan-execute.ts";
-import { resolveStatusMemoryStatusSnapshot } from "./status.scan-memory.ts";
 import { collectStatusScanOverview } from "./status.scan-overview.ts";
 import type { StatusScanResult } from "./status.scan-result.ts";
-import { scanStatusJsonWithPolicy } from "./status.scan.fast-json.js";
 
-/** Runs the status scan for text or JSON command modes. */
+/** Runs the text status scan. */
 export async function scanStatus(
   opts: {
-    json?: boolean;
     timeoutMs?: number;
-    all?: boolean;
     deep?: boolean;
   },
-  _runtime: RuntimeEnv,
 ): Promise<StatusScanResult> {
-  if (opts.json) {
-    // JSON mode uses a policy wrapper so tests and `status-json` can tune fast-path behavior.
-    return await scanStatusJsonWithPolicy(
-      {
-        timeoutMs: opts.timeoutMs,
-        all: opts.all,
-      },
-      _runtime,
-      {
-        commandName: "status --json",
-        resolveHasConfiguredChannels: (cfg, sourceConfig) =>
-          hasConfiguredChannelsForReadOnlyScope({
-            config: cfg,
-            activationSourceConfig: sourceConfig,
-          }),
-        resolveMemory: async ({ cfg, agentStatus, memoryPlugin }) =>
-          await resolveStatusMemoryStatusSnapshot({
-            cfg,
-            agentStatus,
-            memoryPlugin,
-          }),
-      },
-    );
-  }
   return await withProgress(
     {
       label: "Scanning status…",
-      total: 10,
+      total: 9,
       enabled: true,
     },
     async (progress) => {
-      const isFullScan = opts.all === true || opts.deep === true;
+      const isDetailedScan = opts.deep === true;
       const overview = await collectStatusScanOverview({
         env: process.env,
         commandName: "status",
         opts,
         showSecrets: process.env.OPENCLAW_SHOW_SECRETS?.trim() !== "0",
-        includeLiveChannelStatus: isFullScan,
-        includeChannelSetupRuntimeFallback: isFullScan,
-        fetchGitUpdate: isFullScan,
-        includeRegistryUpdate: isFullScan,
+        includeLiveChannelStatus: isDetailedScan,
+        includeChannelSetupRuntimeFallback: isDetailedScan,
+        fetchGitUpdate: isDetailedScan,
+        includeRegistryUpdate: isDetailedScan,
         includeAdvertisedControlUiLinks: true,
         progress,
         labels: {
@@ -74,29 +43,13 @@ export async function scanStatus(
         },
       });
 
-      progress.setLabel("Checking plugins…");
-      const pluginCompatibility = opts.all
-        ? await import("../plugins/status.js").then(({ buildPluginCompatibilitySnapshotNotices }) =>
-            buildPluginCompatibilitySnapshotNotices({ config: overview.cfg }),
-          )
-        : [];
-      progress.tick();
-
       progress.setLabel("Checking memory and sessions…");
       const result = await executeStatusScanFromOverview({
         overview,
-        resolveMemory: async ({ cfg, agentStatus, memoryPlugin }) =>
-          // Memory plugin probing can touch disk/plugin state; reserve it for full scans.
-          opts.all
-            ? await resolveStatusMemoryStatusSnapshot({
-                cfg,
-                agentStatus,
-                memoryPlugin,
-              })
-            : null,
+        resolveMemory: async () => null,
         channelIssues: overview.channelIssues,
         channels: overview.channels,
-        pluginCompatibility,
+        pluginCompatibility: [],
       });
       progress.tick();
